+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## unreleased
+* Add `chargeback_protection_level` to `DisputeSearch`
 * Add webhook sample for `GrantedPaymentMethodRevoked`
 
 ## 4.9.0